Overhaul amcheck documentation
Improve descriptions of the problems that amcheck is most likely to
detect in practice, emphasizing the common cases.  Improve the
description of steps that should be taken in the event of amcheck
finding an issue.  Add commentary on the overhead of running amcheck on
all indexes on a busy production system.