Normalize building 9.6 packages
Previously in af8b090 and 55eae4b there was some thrashing done to be
able to build packages for a postgres server version that was not served
by postgres-server-dev-all package.

This commit removes that thrashing and updates the required
postgres-server-dev-all package version to latest, which allows building
of extensions for 9.6 without gymnastics.