Rename extension to amcheck_next
amcheck_next is a distinct extension to contrib/amcheck for PostgreSQL
extension purposes.  It has non-conflicting symbol names.  This is
preparation for Debian PGDG apt packaging.

This establishes the convention that extension version will not be equal
to package/release version, in order to ease shipping new versions that
only change C code (versions that require no SQL script changes).  We
now follow the style of postgresql-unit: we use simple integers for SQL
extension versions, but traditional mulipart versions for packages.
This convention is intended to make it clear that SQL extension version
and package version are orthogonal.

The first SQL extension version is 1.  The package version is 1.0.

In passing, Change Debian debian/source/format to "1.0", and remove
Travis support.

Per feedback from Christoph Berg.