Add additional enhancements from PostgreSQL 11.
This commit brings amcheck_next in line with contrib/amcheck from
PostgreSQL v11.  There is no change in the user interface, though two
new checks are added:

* Functionality from the recent "Add missing and dangling downlink
checks to amcheck" upstream commit is added.  This is only used when
bt_index_parent_check() is called, and only when heapallindexed
verification is used.

* A new ItemIdGetLength()/IndexTupleSize() crosscheck is added, which
helps with torn-page detection.  This is used when verifying any nbtree
index.